requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
pytest==7.4.3
pytest-flask==1.3.0
celery==5.3.6
redis==5.0.1
boto3==1.34.14
